+++ api/rust/include/LIEF/rust/COFF/Section.hpp
@@ -16,6 +16,7 @@
 #include "LIEF/COFF/Section.hpp"
 #include "LIEF/rust/Abstract/Section.hpp"
 #include "LIEF/rust/COFF/Relocation.hpp"
+#include "LIEF/rust/COFF/String.hpp"
 
 #include "LIEF/rust/Iterator.hpp"
 
@@ -80,6 +81,10 @@ class COFF_Section : public AbstractSection {
 
   auto comdat_info() const {
     return details::try_unique<COFF_Section_ComdataInfo>(impl().comdat_info());
+  }
+
+  auto coff_string() const {
+    return details::try_unique<COFF_String>(impl().coff_string());
   }
 
   auto to_string() const {